Revert "refactor: moved component and its composables to a separate folder"
This reverts commit 4b9e1816fba1e38cddeb8a58180c2b8e23214816.
